Add files via upload
A useful matrix for storing information about a game building. Using the method AddToIndex(), you can add an element to a specific matrix index. If the element goes outside the matrix, it will automatically expand to the required size. It should be noted that the index can be specified as negative, in which case the element will be added to the left or top of the matrix.